Now the program is able o handle numerical labels as it split lines before separating classes. Added some extra doc (examples), the project's structure has been furtherly improved.
